Move config inside pymafia
